refactor collection extensions
